requirements.txt
Django==2.2.3
django-colorful==1.3
django-macaddress==1.6.0
getmac==0.8.1
gunicorn==19.9.0
netaddr==0.7.19
pytz==2019.1
sqlparse==0.3.0
wakeonlan==1.1.6
